refactor(WinningNumbers): 에러enum 파일명 변경 및 null 가드 추가
