+++ src/main/java/lotto/entity/WinningNumbers.java
@@ -1,7 +1,8 @@
 package lotto.entity;
 
-import static lotto.exception.WinnerNumberExceptionMessage.BONUS_NUMBER_DUPLICATE;
-import static lotto.exception.WinnerNumberExceptionMessage.BONUS_NUMBER_OUT_OF_RANGE;
+import static lotto.exception.WinningNumbersExceptionMessage.BONUS_NUMBER_DUPLICATE;
+import static lotto.exception.WinningNumbersExceptionMessage.BONUS_NUMBER_OUT_OF_RANGE;
+import static lotto.exception.WinningNumbersExceptionMessage.NULL_NUMBERS;
 
 import java.util.List;
 import lotto.configuration.LottoConfiguration;
@@ -18,8 +19,11 @@ public class WinningNumbers {
     }
 
     private void validate(List<Integer> mainNumbers, int bonusNumber) {
+        if (mainNumbers == null) {
+            throw new LottoValidationException(NULL_NUMBERS);
+        }
         if (!(LottoConfiguration.LOTTO_MIN_NUMBER.getValue() <= bonusNumber
-                && bonusNumber <= LottoConfiguration.LOTTO_MAX_NUMBER.getValue())) {
+              && bonusNumber <= LottoConfiguration.LOTTO_MAX_NUMBER.getValue())) {
             throw new LottoValidationException(BONUS_NUMBER_OUT_OF_RANGE);
         }
         if (mainNumbers.stream().anyMatch(number -> number == bonusNumber)) {
